requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
akracer==0.0.14
akshare==1.18.49
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
clarabel==0.11.1
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.6
curl_cffi==0.14.0
cvxpy==1.8.2
cycler==0.12.1
databricks-sdk==0.102.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
dnspython==2.8.0
docker==7.1.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.135.2
fastjsonschema==2.21.2
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.3.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
highspy==1.13.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpath==0.82.2
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.4
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
osqp==1.1.1
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-mini-racer==0.6.0
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pymongo==4.16.0
pyparsing==3.3.2
pyqlib==0.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
python-redis-lock==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.1
scs==3.2.11
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==82.0.1
six==1.17.0
skops==0.13.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==1.0.0
tabulate==0.10.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.7
widgetsnbextension==4.0.15
xlrd==2.0.2
yarl==1.23.0
zipp==3.23.0
